-Buy Land function worked -Get $2000 added (add to board) -Add bevel border -Fixed roll dice (Done) -Added Jail (Done) -Board Edit(Done) -Removed Back Function (GUI Bug) -Pay to player function (Done)

+++ src/MonopolyController.java
@@ -37,7 +37,7 @@ public class MonopolyController {
         view.showBuyOption(price);
     }
 
-    public void showPayNotify(int price, int ownership){
+    public void showPayNotify(double price, int ownership){
         view.showPayNotify(price, ownership);
     }
     public void buyLand(){
